Move LatencyFilter to examples
